added subproduct sale support and creation etc

+++ lib/src/models/expense.dart
@@ -10,7 +10,7 @@ class Expense {
   });
 
   String? header;
-  int? amount;
+  double? amount;
   String? description;
   String? modeOfPayment;
   String? id;
@@ -19,7 +19,7 @@ class Expense {
 
   factory Expense.fromMap(Map<String, dynamic> json) => Expense(
         header: json["header"],
-        amount: json["amount"],
+        amount: json["amount"].toDouble(),
         description: json["description"],
         modeOfPayment: json["modeOfPayment"],
         id: json["_id"],
